requirements.txt
# filepath: c:\Users\Marufjon\InsultMedAI\requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
scikit-learn>=1.2.2
pandas>=2.0.0
numpy>=1.24.3
xgboost>=1.7.5
sqlalchemy>=2.0.9
reportlab>=4.0.4
openai>=1.0.0  # Using OpenAI's latest API version
pydantic>=2.0.0
python-dotenv>=1.0.0  # For environment variables
python-multipart>=0.0.6  # For file uploads
requests>=2.30.0  # For API requests
pydantic-settings>=2.0.0  # For settings management
joblib>=1.2.0  # For model loading
starlette>=0.27.0  # For FastAPI middleware
# Auth dependencies
python-jose[cryptography]>=3.3.0  # For JWT tokens
passlib[bcrypt]>=1.7.4  # For password hashing
email-validator>=2.0.0  # For email validation
google-auth>=2.22.0  # For Google OAuth2
google-auth-oauthlib>=1.0.0  # For Google OAuth2 integration
requests-oauthlib>=1.3.1  # For OAuth2 requests
alembic>=1.12.0  # For database migrations
